refactor: add test config providers to TestConfig class
