+++ backend/src/TestConfigurationAccessor/ConfigProviders/AppConfigurationTestConfigProvider.cs
@@ -1,0 +1,26 @@
+﻿using Microsoft.Extensions.Configuration;
+
+namespace TestConfigurationAccessor.ConfigProviders
+{
+    internal class AppConfigurationTestConfigProvider : ITestConfigProvider
+    {
+        private const string AppConfigPrefix = "appcfg";
+
+
+        public bool AddConfiguration(ConfigurationManager builder, string envName, string? prefix, string? suffix)
+        {
+            if (prefix == null || suffix == null || prefix != AppConfigPrefix)
+            {
+                return false;
+            }
+
+
+            builder.AddAzureAppConfiguration(cfg =>
+            {
+                cfg.Connect("");
+            });
+
+            return true;
+        }
+    }
+}
