+++ backend/src/TestConfigurationAccessor/ConfigProviders/ITestConfigProvider.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Extensions.Configuration;
+
+namespace TestConfigurationAccessor.ConfigProviders
+{
+    internal interface ITestConfigProvider
+    {
+        bool AddConfiguration(ConfigurationManager builder, string envName, string? prefix, string? suffix);
+    }
+}
