+++ backend/src/TestConfigurationAccessor/TestConfig.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Configuration;
+using TestConfigurationAccessor.ConfigProviders;
 
 namespace TestConfigurationAccessor
 {
@@ -6,14 +7,23 @@ namespace TestConfigurationAccessor
     /// Provides access to test configuration from provider selected by env name. 
     /// Source of configuration is obtained by checking "ENV" environment variable.
     /// If the environment variable contains prefix (ENV="exampleprefix-suffix") then source of configuration is obtained using provider configured for this prefix.
-    /// Default value of "ENV" is "local".
+    /// Suffix usage depends on type of provider.
+    /// Default value of "ENV" is "local" which results in usage of default file-based provider. The file-based provider uses ENV variable to choose source file of settings.
+    /// For example: 
+    /// - if ENV="local", provider loads: settings.json, settings.local.json
+    /// - if ENV="docker", provider loads: settings.json, settings.docker.json
     /// </summary>
     public static class TestConfig
     {
         private const string LocalEnvName = "local";
-        private const string AppConfigPrefix = "appcfg";
         private static Lazy<IConfigurationRoot> _instance = new Lazy<IConfigurationRoot>(BuildConfiguration);
-        
+
+        private static readonly IEnumerable<ITestConfigProvider> TestConfigProviders = new ITestConfigProvider[]
+        {
+            new FileTestConfigProvider(),
+            new AppConfigurationTestConfigProvider()
+        };
+
         /// <summary>
         /// Lazy initialized configuration instance
         /// </summary>
@@ -23,41 +33,19 @@ namespace TestConfigurationAccessor
         private static IConfigurationRoot BuildConfiguration()
         {
             var envName = GetEnvironmentName();
-
             var cfgManager = new ConfigurationManager();
-
             cfgManager.AddUserSecrets(typeof(TestConfig).Assembly);
 
             var (prefix, suffix) = (envName.Split('-').Length > 0 ? envName.Split('-')[0] : null, envName.Split('-').Length > 1 ? envName.Split('-')[1] : null);
-            if (prefix != null && suffix != null && prefix == (AppConfigPrefix))
+            foreach (var testConfigProvider in TestConfigProviders)
             {
-                AddAppConfig(cfgManager, "", suffix);
-            }
-            else
-            {
-                AddDefaultJsonFile(cfgManager);
-                AddJsonFile(cfgManager, envName);
+                if (testConfigProvider.AddConfiguration(cfgManager, envName, prefix, suffix))
+                {
+                    break;
+                }
             }
 
             return cfgManager;
-        }
-
-        private static void AddDefaultJsonFile(IConfigurationBuilder builder)
-        {
-            builder.AddJsonFile("settings.json", optional: true);
-        }
-
-        private static void AddJsonFile(IConfigurationBuilder builder, string envName)
-        {
-            builder.AddJsonFile($"settings.{envName}.json");
-        }
-
-        private static void AddAppConfig(IConfigurationBuilder builder, string connectionString, string envName)
-        {
-            builder.AddAzureAppConfiguration(cfg =>
-            {
-                cfg.Connect(connectionString);
-            });
         }
 
         private static string GetEnvironmentName()
